feat: setup evm-account-kit
