+++ packages/boot/test/orchestration/snapshots/orchestration.test.ts.md
@@ -71,7 +71,10 @@ Generated by [AVA](https://avajs.dev).
         }
         'published.agoricNames.chain.axelar',
         {
+          bech32Prefix: 'axelar',
           chainId: 'axelar',
+          namespace: 'cosmos',
+          reference: 'axelar',
           stakingTokens: [
             {
               denom: 'uaxl',
