+++ packages/orchestration/test/examples/snapshots/send-anywhere.test.ts.md
@@ -38,6 +38,7 @@ Generated by [AVA](https://avajs.dev).
         bech32PrefixToChainName: {
           agoric: 'agoric',
           archway: 'archway',
+          axelar: 'axelar',
           bze: 'beezee',
           celestia: 'celestia',
           core: 'coreum',
@@ -153,10 +154,13 @@ Generated by [AVA](https://avajs.dev).
             ],
           },
           axelar: {
+            bech32Prefix: 'axelar',
             chainId: 'axelar',
             icaEnabled: false,
             icqEnabled: false,
+            namespace: 'cosmos',
             pfmEnabled: false,
+            reference: 'axelar',
             stakingTokens: [
               {
                 denom: 'uaxl',
