+++ packages/orchestration/test/snapshots/chain-info.test.js.md
@@ -68,7 +68,10 @@ Generated by [AVA](https://avajs.dev).
         }
         'axelar',
         {
+          bech32Prefix: 'axelar',
           chainId: 'axelar',
+          namespace: 'cosmos',
+          reference: 'axelar',
           stakingTokens: [
             {
               denom: 'uaxl',
